Fix deprecated listener in Symfony < 4.3
